requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
fastapi-pagination==0.11.2
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
MarkupSafe==2.1.2
msgpack==1.0.4
packaging==23.0
pluggy==1.0.0
priority==2.0.0
psutil==5.9.4
psycopg==3.1.8
pydantic==1.10.4
PyJWT==2.6.0
pytest==7.2.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3986==1.5.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
sqladmin==0.8.0
SQLAlchemy==1.4.46
starlette==0.22.0
toml==0.10.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
zope.event==4.6
zope.interface==5.5.2
